requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.19.0
google-generativeai==0.8.5
python-dotenv==1.0.0
beautifulsoup4==4.12.0
requests==2.31.0
schedule==1.2.0
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.13.0
pillow==10.2.0
openpyxl==3.1.2
aiohttp==3.9.3
asyncio-mqtt==0.16.2
python-dateutil==2.8.2
pytz==2024.1
pyyaml==6.0.1
jsonschema==4.21.1
urllib3==2.2.0
certifi==2024.2.2
scipy>=1.11.0
scikit-learn>=1.3.0
networkx>=3.1